# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/projects/{projectID}/databases/{databaseID}/tables/{tableID}/export:
    get:
      operationId: exportDatabaseRows
      description: Stream a permission-filtered table snapshot as JSON or CSV. JSON preserves the DatabaseRow shape; CSV uses JSON literals for declared data columns. Exports are bounded by the limit query parameter.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }, { projectAppCookie: [] }, {}]
      parameters:
        - { $ref: '#/components/parameters/ProjectID' }
        - { $ref: '#/components/parameters/DatabaseID' }
        - { $ref: '#/components/parameters/TableID' }
        - { name: format, in: query, schema: { type: string, enum: [json, csv], default: json } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 10000, default: 1000 } }
      responses:
        '200': { description: Permission-filtered table export, content: { application/json: { schema: { $ref: '#/components/schemas/DatabaseRowsExport' } }, text/csv: { schema: { type: string, format: binary } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/projects/{projectID}/databases/{databaseID}/tables/{tableID}/rows/transaction:
    post:
      operationId: transactDatabaseRows
      description: Atomically apply up to 100 creates, updates, and deletes to one table. Every operation is permission-checked and relationship-checked; any failure rolls the entire batch back.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }, { projectAppCookie: [] }, {}]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/DatabaseID' }, { $ref: '#/components/parameters/TableID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DatabaseRowsTransactionRequest' } } } }
      responses:
        '200': { description: Atomic row transaction result, content: { application/json: { schema: { $ref: '#/components/schemas/DatabaseRowsTransactionResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes include users, databases, storage, functions, sites, webhooks, realtime.read, and messaging read/write pairs; never send this header through the Console browser bridge. }
  parameters:
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
    DatabaseID: { name: databaseID, in: path, required: true, schema: { type: string, format: uuid } }
...
    TableID: { name: tableID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    RateLimited: { description: Public Auth rate limit exceeded, headers: { Retry-After: { description: Seconds until another attempt may be made, schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ServiceUnavailable: { description: Public Auth protection or a required dependency is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    DatabasePermission: { type: string, pattern: '^(any|users|user:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$', description: Canonical application grant. Only any, users, and user:<project-user-uuid> are supported. }
...
    DatabaseRow: { type: object, additionalProperties: false, required: [id, table_id, project_id, data, read_permissions, update_permissions, delete_permissions, created_at, updated_at], properties: { id: { type: string, format: uuid }, table_id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, data: { type: object, description: Typed user columns; unknown/system keys are rejected. }, read_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } }, update_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } }, delete_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } }, creator_project_user_id: { type: [string, 'null'], format: uuid }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
...
    DatabaseRowsTransactionOperation: { type: object, additionalProperties: false, required: [action], properties: { action: { type: string, enum: [create, update, delete] }, id: { type: string, format: uuid, description: Required for update/delete; optional stable id for create. }, data: { type: object, description: Typed data for create or partial data for update. }, read_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } }, update_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } }, delete_permissions: { type: array, items: { $ref: '#/components/schemas/DatabasePermission' } } } }
    DatabaseRowsTransactionRequest: { type: object, additionalProperties: false, required: [operations], properties: { operations: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/DatabaseRowsTransactionOperation' } } } }
...
    DatabaseRowsExport: { type: object, additionalProperties: false, required: [rows, count], properties: { rows: { type: array, items: { $ref: '#/components/schemas/DatabaseRow' } }, count: { type: integer, minimum: 0, maximum: 10000 } } }
...
    DatabaseRowsTransactionResponse: { type: object, additionalProperties: false, required: [rows, deleted_ids, count], properties: { rows: { type: array, items: { $ref: '#/components/schemas/DatabaseRow' } }, deleted_ids: { type: array, items: { type: string, format: uuid } }, count: { type: integer, minimum: 1, maximum: 100 } } }